페이지 추가를 위한  AppPage 코드 수정
