+++ src/main/frontend/src/board/BoardWrite.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function BoardWrite() {
+
+    return (
+        <div>
+            <h1>boardWrite</h1>
+        </div>
+    )
+}
+
+export default BoardWrite;
